Rename InputTriggerVerification to InputSelectClaim
